requirements.txt
annotated-types==0.7.0
argcomplete==3.5.0
audioread==3.0.1
blinker==1.8.2
boto3==1.35.10
botocore==1.35.10
certifi==2024.8.30
cffi==1.17.0
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
durationpy==0.7
ecdsa==0.19.0
Flask==3.0.3
hjson==3.1.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kappa==0.6.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
msgpack==1.0.8
numba==0.60.0
numpy==2.0.2
packaging==24.1
placebo==0.9.0
platformdirs==4.2.2
pooch==1.8.2
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.1
setuptools==74.1.0
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
text-unidecode==1.3
threadpoolctl==3.5.0
toml==0.10.2
tqdm==4.66.5
troposphere==4.8.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wheel==0.44.0
zappa==0.59.0
